Fix FREEBIND_RANDOM environment variable getting modified by the initialisation, breaking it for child processes
strtok_r replaces delimiters with NUL bytes. Because getenv returns a pointer to the value in the environment, this truncates the value at the first comma or slash, effectively breaking freebind for any child process.

Also fix the incorrect usage of strtok_r: the first argument must be NULL on calls after the first.